Added MenuComponent
Added:

Menu bar that slides automatically to the top based on the scroll position of the page with the header.

MenuButton to the MenuComponent that hrefs to a link given on the attributes

Dynamic generation of the status containers based on the Enums of the TaskStatus in the backend.

Header Content with a logo, name and version number